+++ CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#4728](https://github.com/open-telemetry/opentelemetry-python/pull/4728))
 - Permit to override default HTTP OTLP exporters headers
   ([#4634](https://github.com/open-telemetry/opentelemetry-python/pull/4634))
+- semantic-conventions: Bump to 1.37.0
+  ([#4731](https://github.com/open-telemetry/opentelemetry-python/pull/4731))
 
 ## Version 1.36.0/0.57b0 (2025-07-29)